create the Encoder/Decoder interfaces and implementations
Create the Encoder/Decoder interfaces and implementations to help
with tests and provide a systematic way to transform any graph input to
any other.

The goal is that all the tests that requires some input graph file use
the Simple format.

Maybe in the future I should move the encoding package outside of
internal, I think this package would be useful not only for tests but to
the end user too.